Update doc slightly, add some more thoughts to the code regarding sending messages between tasks
Starting to think that we don't even need queues. If we acquire and transmit data every 2 ms, then we don't even need queues because we can just take advantage of synchronization mechanisms and task priorities to get the data out. This would greatly simplify things and might be a good first step for a data readout