Updates PC-side data to store packets in a text file and added PWM for magnets
PC-side program
- Logs data (either raw or interpreted, whatever) to a text file
- Text file buffer is flushed every 1000 receptions to improve reliability (encourage disk writes)

In tim.c/.h, a function was created to interface with the H-bridges for the magnets. COAST, POSITIVECURRENT, and NEGATIVECURRENT settings were tested using 4 LEDs on a breadboard. These all work. However, BRAKE is not working properly, and although COAST produces the desired effect, it's happening for the wrong reason. See the TODO in tim.c for more details