Fixed some problems with the queues transferring information between tasks. Also, temperature sensor task now works in a very rudimentary manner that can be built off of
In freertos.c, data was actually not being passed through the queues correctly. The updated method uses structs to pass messages and I have verified that it works correctly.

The current method for sensing temperature is very simple and I hacked it in just so that I could get SOME level of ADC working that we could (potentially) demo.