Correct all the RTC stuff in initialization. Should be working now.
The last step in initialization is making it so that the MCU can be started before the PC application. I.e., the microcontroller will either enter a blocked state until it received a start command from the PC, OR we just protect the transmissions with checksums and the PC doesn't send ACK unless the checksum matches. Could try using the CRC32 module for this.